# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportBadgeDefinition": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "credit_value": {
            "title": "Credit Value",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "display_order": {
            "title": "Display Order",
            "type": "integer"
          },
          "icon_key": {
            "title": "Icon Key",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "threshold": {
            "title": "Threshold",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "description",
          "category",
          "threshold",
          "icon_key",
          "credit_value",
          "display_order",
          "active"
        ],
        "title": "ExportBadgeDefinition",
        "type": "object"
      },
      "ExportBook": {
        "properties": {
          "authors": {
            "items": {
              "type": "string"
            },
            "title": "Authors",
            "type": "array"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "external_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Source Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isbn_10": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 10"
          },
          "isbn_13": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Isbn 13"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "metadata_source": {
            "$ref": "#/components/schemas/MetadataSource"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "published_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published Date"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "subtitle",
          "authors",
          "isbn_10",
          "isbn_13",
          "cover_url",
          "publisher",
          "published_date",
          "page_count",
          "description",
          "language",
          "metadata_source",
          "external_source_id",
          "created_at",
          "updated_at"
        ],
        "title": "ExportBook",
        "type": "object"
      },
      "ExportReader": {
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "avatar_key",
          "created_at",
          "updated_at"
        ],
        "title": "ExportReader",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MetadataSource": {
        "enum": [
          "manual",
          "google_books",
          "open_library"
        ],
        "title": "MetadataSource",
        "type": "string"
      },
...
      "ReaderCreate": {
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "pattern": "^[a-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "name": {
            "maxLength": 80,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ReaderCreate",
        "type": "object"
      },
      "ReaderLoginInvitationCreate": {
        "properties": {
          "email": {
            "title": "Email",
            "type": "string"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          }
        },
        "required": [
          "reader_id",
          "email"
        ],
        "title": "ReaderLoginInvitationCreate",
        "type": "object"
      },
      "ReaderLoginInvitationResponse": {
        "properties": {
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reader_id",
          "email",
          "accepted"
        ],
        "title": "ReaderLoginInvitationResponse",
        "type": "object"
      },
      "ReaderResponse": {
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "household_id": {
            "format": "uuid",
            "title": "Household Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "household_id",
          "name",
          "avatar_key",
          "created_at",
          "updated_at"
        ],
        "title": "ReaderResponse",
        "type": "object"
      },
      "ReaderUpdate": {
        "properties": {
          "avatar_key": {
            "anyOf": [
              {
                "maxLength": 40,
                "minLength": 1,
                "pattern": "^[a-z0-9_-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Key"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 80,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "ReaderUpdate",
        "type": "object"
      },
      "ReadingDataExport": {
        "properties": {
          "badge_definitions": {
            "items": {
              "$ref": "#/components/schemas/ExportBadgeDefinition"
            },
            "title": "Badge Definitions",
            "type": "array"
          },
          "books": {
            "items": {
              "$ref": "#/components/schemas/ExportBook"
            },
            "title": "Books",
            "type": "array"
          },
          "exported_at": {
            "format": "date-time",
            "title": "Exported At",
            "type": "string"
          },
          "household_id": {
            "format": "uuid",
            "title": "Household Id",
            "type": "string"
          },
          "household_name": {
            "title": "Household Name",
            "type": "string"
          },
          "reader_badges": {
            "items": {
              "$ref": "#/components/schemas/ExportReaderBadge"
            },
            "title": "Reader Badges",
            "type": "array"
          },
          "reader_books": {
            "items": {
              "$ref": "#/components/schemas/ExportReaderBook"
            },
            "title": "Reader Books",
            "type": "array"
          },
          "readers": {
            "items": {
              "$ref": "#/components/schemas/ExportReader"
            },
            "title": "Readers",
            "type": "array"
          },
          "reading_sessions": {
            "items": {
              "$ref": "#/components/schemas/ExportReadingSession"
            },
            "title": "Reading Sessions",
            "type": "array"
          },
          "reward_items": {
            "items": {
              "$ref": "#/components/schemas/ExportRewardItem"
            },
            "title": "Reward Items",
            "type": "array"
          },
          "reward_redemptions": {
            "items": {
              "$ref": "#/components/schemas/ExportRewardRedemption"
            },
            "title": "Reward Redemptions",
            "type": "array"
          },
          "reward_transactions": {
            "items": {
              "$ref": "#/components/schemas/ExportRewardTransaction"
            },
            "title": "Reward Transactions",
            "type": "array"
          },
          "schema_version": {
            "default": 2,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "exported_at",
          "household_id",
          "household_name",
          "readers",
          "books",
          "reader_books",
          "reading_sessions",
          "badge_definitions",
          "reader_badges",
          "reward_items",
          "reward_transactions",
          "reward_redemptions"
        ],
        "title": "ReadingDataExport",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "My Reading Tracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exports/reading-data": {
      "get": {
        "operationId": "export_reading_data_api_v1_exports_reading_data_get",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "json",
              "enum": [
                "json",
                "csv",
                "finished-books-csv"
              ],
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadingDataExport"
                }
              },
              "text/csv": {}
            },
            "description": "Complete JSON backup, reading-session CSV, or finished-books CSV"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export Reading Data",
        "tags": [
          "exports"
        ]
      }
    },
    "/api/v1/exports/school-reading-report": {
      "get": {
        "operationId": "export_school_reading_report_api_v1_exports_school_reading_report_get",
        "parameters": [
          {
            "in": "query",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Date To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/pdf": {}
            },
            "description": "Printable school reading achievement report"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export School Reading Report",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/api/v1/reader-login-invitations": {
      "get": {
        "operationId": "list_invitations_api_v1_reader_login_invitations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReaderLoginInvitationResponse"
                  },
                  "title": "Response List Invitations Api V1 Reader Login Invitations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Invitations",
        "tags": [
          "reader access"
        ]
      },
      "post": {
        "operationId": "create_invitation_api_v1_reader_login_invitations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderLoginInvitationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderLoginInvitationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Invitation",
        "tags": [
          "reader access"
        ]
      }
    },
    "/api/v1/reader-login-invitations/{invitation_id}": {
      "delete": {
        "operationId": "delete_invitation_api_v1_reader_login_invitations__invitation_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "invitation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Invitation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Invitation",
        "tags": [
          "reader access"
        ]
      }
    },
    "/api/v1/readers": {
      "get": {
        "operationId": "list_readers_api_v1_readers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReaderResponse"
                  },
                  "title": "Response List Readers Api V1 Readers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Readers",
        "tags": [
          "readers"
        ]
      },
      "post": {
        "operationId": "create_reader_api_v1_readers_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reader",
        "tags": [
          "readers"
        ]
      }
    },
    "/api/v1/readers/{reader_id}": {
      "delete": {
        "operationId": "delete_reader_api_v1_readers__reader_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "confirm_history",
            "required": false,
            "schema": {
              "default": false,
              "title": "Confirm History",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Reader",
        "tags": [
          "readers"
        ]
      },
      "get": {
        "operationId": "get_reader_api_v1_readers__reader_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Reader",
        "tags": [
          "readers"
        ]
      },
      "patch": {
        "operationId": "update_reader_api_v1_readers__reader_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReaderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Reader",
        "tags": [
          "readers"
        ]
      }
    },
...
  }
}